+++ Model/FormRepository.php
@@ -20,7 +20,11 @@ class FormRepository
     /**
      * @var array
      */
-    protected $loadedForms = [];
+    protected $loadedFormsByIds = [];
+    /**
+     * @var array
+     */
+    protected $loadedFormsByCodes = [];
 
     /**
      * FormRepository constructor.
@@ -39,7 +43,8 @@ class FormRepository
      */
     public function getById($formId, $storeId = null, $graceful = false)
     {
-        if (!isset($this->loadedForms[$formId])) {
+        $storeKey = $storeId ?? 'null';
+        if (!isset($this->loadedFormsByIds[$formId][$storeKey])) {
             $form = $this->formFactory->create();
             $form->setStoreId($storeId);
             $form->getResource()->load($form, $formId);
@@ -50,8 +55,46 @@ class FormRepository
                     throw new NoSuchEntityException(__('Form with id "%1" does not exist.', $formId));
                 }
             }
-            $this->loadedForms[$formId] = $form;
+            $this->addFormToLoaded($form, $storeKey);
         }
-        return $this->loadedForms[$formId];
+
+        return $this->loadedFormsByIds[$formId][$storeKey];
+    }
+
+    /**
+     * @param $formId
+     * @param null $storeId
+     * @param false $graceful
+     */
+    public function getByCode($formCode, $storeId = null, $graceful = true)
+    {
+        $storeKey = $storeId ?? 'null';
+        if (!isset($this->loadedFormsByCodes[$formCode])) {
+            $form = $this->formFactory->create();
+            $form->setStoreId($storeId);
+            $form->getResource()->load($form, $formCode, 'form_code');
+            if (!$form->getId()) {
+                if ($graceful) {
+                    return $form;
+                } else {
+                    throw new NoSuchEntityException(__('Form with code "%1" does not exist.', $formCode));
+                }
+            }
+            $this->addFormToLoaded($form, $storeKey);
+        }
+
+        $formId = $this->loadedFormsByCodes[$formCode];
+        return $this->loadedFormsByIds[$formId][$storeKey];
+    }
+
+    /**
+     * @param $form
+     */
+    protected function addFormToLoaded($form, $storeKey = 'null')
+    {
+        $this->loadedFormsByIds[$form->getId()][$storeKey] = $form;
+        if ($form->getFormCode()) {
+            $this->loadedFormsByCodes[$form->getFormCode()] = $form->getId();
+        }
     }
 }